Handle credentials so allow private cloning
Rather than using the same mechanism used by checkout to handle libgit2
callbacks, Swift 2 now lets us pass through Swift function as C function
pointers.

When using a Swift function as a function pointer, it cannot be a
closure. This means that we have to wrap up the CredentialProvider and
store it in the payload so we can use it inside of the callback
function.

This commit implements a single credential type, SSH in-memory, as this
allows us to test the clone function.

The test requires you passing through a set of environment variables
otherwise it will not run. This is so that secret/private information
isn't leaked into the repository. The required variables are as follows:

   - SG2TestPrivateRepo - the url of the private jrepo to clone;
   - SG2TestUsername - the url of the user that will be connecting;
   - SG2TestPublicKey - the public key data to be used;
   - SG2TestPrivateKey - the private key data;
   - SG2TestPassphrase - passphrase needed to use the private key (blank
     if none).